+++ src/connection_settings.h
@@ -19,6 +19,8 @@ struct ConnectionSettings
 
   bool no_video = false;
   bool no_audio = false;
+  bool force_i420 = false;
+  bool use_native = false;
   std::string video_codec = "VP8";
   std::string audio_codec = "OPUS";
   int video_bitrate = 0;
@@ -34,11 +36,12 @@ struct ConnectionSettings
   bool sora_auto_connect = false;
   nlohmann::json sora_metadata;
 
-  std::string p2p_document_root;
+  std::string test_document_root;
 
   std::string ayame_signaling_host;
-  std::string ayame_client_id;
   std::string ayame_room_id;
+  std::string ayame_client_id;
+  std::string ayame_signaling_key = "";
 
   int getWidth() {
     if (resolution == "QVGA") {
@@ -89,13 +92,13 @@ struct ConnectionSettings
     os << "priority: " << cs.priority << "\n";
     os << "port: " << cs.port << "\n";
     os << "ayame_signaling_host: " << cs.ayame_signaling_host << "\n";
-    os << "ayame_client_id: " << cs.ayame_client_id << "\n";
     os << "ayame_room_id: " << cs.ayame_room_id << "\n";
+    os << "ayame_client_id: " << cs.ayame_client_id << "\n";
     os << "sora_signaling_host: " << cs.sora_signaling_host << "\n";
     os << "sora_channel_id: " << cs.sora_channel_id << "\n";
     os << "sora_auto_connect: " << (cs.sora_auto_connect ? "true" : "false") << "\n";
     os << "sora_metadata: " << cs.sora_metadata << "\n";
-    os << "p2p_document_root: " << cs.p2p_document_root << "\n";
+    os << "test_document_root: " << cs.test_document_root << "\n";
     return os;
   }
 };