+++ src/mac_helper/mac_capturer.h
@@ -15,10 +15,14 @@
 
 #include "api/media_stream_interface.h"
 #include "api/scoped_refptr.h"
+#include "base/RTCMacros.h"
 #include "modules/video_capture/video_capture.h"
 #include "rtc_base/thread.h"
 
 #include "rtc/scalable_track_source.h"
+
+RTC_FWD_DECL_OBJC_CLASS(RTCCameraVideoCapturer);
+RTC_FWD_DECL_OBJC_CLASS(RTCVideoSourceAdapter);
 
 class MacCapturer : public ScalableVideoTrackSource,
                     public rtc::VideoSinkInterface<webrtc::VideoFrame> {
@@ -38,8 +42,8 @@ class MacCapturer : public ScalableVideoTrackSource,
  private:
   void Destroy();
 
-  void* capturer_;
-  void* adapter_;
+  RTCCameraVideoCapturer* capturer_;
+  RTCVideoSourceAdapter* adapter_;
 };
 
 #endif  // TEST_MAC_CAPTURER_H_