+++ src/main.cpp
@@ -19,7 +19,11 @@
 #ifdef __APPLE__
 #include "mac_helper/mac_capturer.h"
 #else
+#if USE_MMAL_ENCODER
+#include "v4l2_video_capturer/v4l2_video_capturer.h"
+#else
 #include "rtc/device_video_capturer.h"
+#endif
 #endif
 #endif
 
@@ -37,12 +41,12 @@ int main(int argc, char* argv[])
   ConnectionSettings cs;
 
   bool is_daemon = false;
-  bool use_p2p = false;
+  bool use_test = false;
   bool use_ayame = false;
   bool use_sora = false;
   int log_level = rtc::LS_NONE;
 
-  Util::parseArgs(argc, argv, is_daemon, use_p2p, use_ayame, use_sora, log_level, cs);
+  Util::parseArgs(argc, argv, is_daemon, use_test, use_ayame, use_sora, log_level, cs);
 
 #ifndef _MSC_VER
   if (is_daemon)
@@ -79,8 +83,13 @@ int main(int argc, char* argv[])
   rtc::scoped_refptr<MacCapturer> capturer =
           MacCapturer::Create(cs.getWidth(), cs.getHeight(), cs.framerate, 0);
 #else
+#if USE_MMAL_ENCODER
+  rtc::scoped_refptr<V4L2VideoCapture> capturer =
+          V4L2VideoCapture::Create(cs);
+#else
   rtc::scoped_refptr<DeviceVideoCapturer> capturer =
           DeviceVideoCapturer::Create(cs.getWidth(), cs.getHeight(), cs.framerate);
+#endif
 #endif
   if (!capturer)
   {
@@ -104,9 +113,9 @@ int main(int argc, char* argv[])
         std::make_shared<SoraServer>(ioc, endpoint, rtc_manager.get(), cs)->run();
       }
 
-      if (use_p2p) {
+      if (use_test) {
         const boost::asio::ip::tcp::endpoint endpoint{boost::asio::ip::make_address("0.0.0.0"), static_cast<unsigned short>(cs.port)};
-        std::make_shared<P2PServer>(ioc, endpoint, std::make_shared<std::string>(cs.p2p_document_root), rtc_manager.get(), cs)->run();
+        std::make_shared<P2PServer>(ioc, endpoint, std::make_shared<std::string>(cs.test_document_root), rtc_manager.get(), cs)->run();
       }
 
       if (use_ayame) {