+++ src/rtc/manager.cpp
@@ -4,9 +4,9 @@
 #include "api/audio_codecs/builtin_audio_decoder_factory.h"
 #include "api/audio_codecs/builtin_audio_encoder_factory.h"
 #include "api/create_peerconnection_factory.h"
-#include "api/task_queue/global_task_queue_factory.h"
+#include "api/task_queue/default_task_queue_factory.h"
 #include "api/video_track_source_proxy.h"
-#include "logging/rtc_event_log/rtc_event_log_factory.h"
+#include "api/rtc_event_log/rtc_event_log_factory.h"
 #include "media/engine/webrtc_media_engine.h"
 #include "modules/audio_device/include/audio_device.h"
 #include "modules/audio_processing/include/audio_processing.h"
@@ -14,6 +14,7 @@
 #include "modules/video_capture/video_capture_factory.h"
 #include "rtc_base/ssl_adapter.h"
 #include "rtc_base/logging.h"
+#include "absl/memory/memory.h"
 
 #include "scalable_track_source.h"
 #include "manager.h"
@@ -32,9 +33,8 @@
 #endif
 
 #if USE_MMAL_ENCODER
+#include "api/video_codecs/video_encoder_factory.h"
 #include "hw_video_encoder_factory.h"
-#include "api/video_codecs/builtin_video_decoder_factory.h"
-#include "absl/memory/memory.h"
 #endif
 
 RTCManager::RTCManager(ConnectionSettings conn_settings,
@@ -60,34 +60,42 @@ RTCManager::RTCManager(ConnectionSettings conn_settings,
     audio_layer = webrtc::AudioDeviceModule::kDummyAudio;
   }
 
-  std::unique_ptr<cricket::MediaEngineInterface> media_engine = cricket::WebRtcMediaEngineFactory::Create(
-#if USE_ROS
-      ROSAudioDeviceModule::Create(_conn_settings, &webrtc::GlobalTaskQueueFactory()),
-#else
-	  webrtc::AudioDeviceModule::Create(audio_layer, &webrtc::GlobalTaskQueueFactory()),
-#endif
-      webrtc::CreateBuiltinAudioEncoderFactory(),
-      webrtc::CreateBuiltinAudioDecoderFactory(),
-#ifdef __APPLE__
-      CreateObjCEncoderFactory(),
-      CreateObjCDecoderFactory(),
-#else
-#if USE_MMAL_ENCODER
-      std::unique_ptr<webrtc::VideoEncoderFactory>(absl::make_unique<HWVideoEncoderFactory>()),
-#else
-      webrtc::CreateBuiltinVideoEncoderFactory(),
-#endif
-      webrtc::CreateBuiltinVideoDecoderFactory(),
-#endif
-      nullptr /* audio_mixer */,
-      webrtc::AudioProcessingBuilder().Create());
   webrtc::PeerConnectionFactoryDependencies dependencies;
   dependencies.network_thread = _networkThread.get();
   dependencies.worker_thread = _workerThread.get();
   dependencies.signaling_thread = _signalingThread.get();
-  dependencies.media_engine = std::move(media_engine);
+  dependencies.task_queue_factory = webrtc::CreateDefaultTaskQueueFactory();
   dependencies.call_factory = webrtc::CreateCallFactory();
-  dependencies.event_log_factory = webrtc::CreateRtcEventLogFactory();
+  dependencies.event_log_factory = absl::make_unique<webrtc::RtcEventLogFactory>(
+      dependencies.task_queue_factory.get());
+
+  // media_dependencies
+  cricket::MediaEngineDependencies media_dependencies;
+  media_dependencies.task_queue_factory = dependencies.task_queue_factory.get();
+#if USE_ROS
+    media_dependencies.adm = ROSAudioDeviceModule::Create(_conn_settings, dependencies.task_queue_factory.get());
+#else
+    media_dependencies.adm = webrtc::AudioDeviceModule::Create(audio_layer, dependencies.task_queue_factory.get());
+#endif
+  media_dependencies.audio_encoder_factory = webrtc::CreateBuiltinAudioEncoderFactory();
+  media_dependencies.audio_decoder_factory = webrtc::CreateBuiltinAudioDecoderFactory();
+#ifdef __APPLE__
+    media_dependencies.video_encoder_factory = CreateObjCEncoderFactory();
+    media_dependencies.video_decoder_factory = CreateObjCDecoderFactory();
+#else
+#if USE_MMAL_ENCODER
+    media_dependencies.video_encoder_factory = std::unique_ptr<webrtc::VideoEncoderFactory>(absl::make_unique<HWVideoEncoderFactory>());
+#else
+    media_dependencies.video_encoder_factory = webrtc::CreateBuiltinVideoEncoderFactory();
+#endif
+    media_dependencies.video_decoder_factory = webrtc::CreateBuiltinVideoDecoderFactory();
+#endif
+  media_dependencies.audio_mixer = nullptr;
+  media_dependencies.audio_processing = webrtc::AudioProcessingBuilder().Create();
+
+  dependencies.media_engine =
+      cricket::CreateMediaEngine(std::move(media_dependencies));
+
   _factory = webrtc::CreateModularPeerConnectionFactory(std::move(dependencies));
   if (!_factory.get())
   {
@@ -135,12 +143,12 @@ std::shared_ptr<RTCConnection> RTCManager::createConnection(
     return nullptr;
   }
 
-  std::string stream_id = Util::generateRundomChars();
+  std::string stream_id = Util::generateRandomChars();
 
   if (!_conn_settings.no_audio)
   {
     rtc::scoped_refptr<webrtc::AudioTrackInterface> audio_track(
-        _factory->CreateAudioTrack(Util::generateRundomChars(),
+        _factory->CreateAudioTrack(Util::generateRandomChars(),
                                    _factory->CreateAudioSource(cricket::AudioOptions())));
     if (audio_track)
     {
@@ -157,7 +165,7 @@ std::shared_ptr<RTCConnection> RTCManager::createConnection(
 
   if (_video_source) {
     rtc::scoped_refptr<webrtc::VideoTrackInterface> video_track(
-            _factory->CreateVideoTrack(Util::generateRundomChars(), _video_source));
+            _factory->CreateVideoTrack(Util::generateRandomChars(), _video_source));
     if (video_track)
     {
       if (_conn_settings.fixed_resolution) {